requirements.txt
anyio==4.11.0
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
cryptography==46.0.3
distlib==0.4.0
filelock==3.20.0
h11==0.16.0
hatch==1.15.1
hatchling==1.27.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jeepney==0.9.0
keyring==25.6.0
markdown-it-py==4.0.0
mdurl==0.1.2
more-itertools==10.8.0
numpy==2.3.4
packaging==25.0
pathspec==0.12.1
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
ptyprocess==0.7.0
pycparser==2.23
Pygments==2.19.2
-e git+https://github.com/lcb-dev/PyMaze.git@3d2ed042b626976b2d8ad3bd836be047405152a2#egg=PyMaze
rich==14.2.0
SecretStorage==3.4.0
shellingham==1.5.4
sniffio==1.3.1
tomli_w==1.2.0
tomlkit==0.13.3
trove-classifiers==2025.9.11.17
userpath==1.9.2
uv==0.9.7
virtualenv==20.35.4
zstandard==0.25.0
